+++ packages/contracts/source/events.ts
@@ -84,3 +84,11 @@ export enum QueueEvent {
 	Finished = "queue.finished",
 	Failed = "queue.failed",
 }
+
+export enum WebhookEvent {
+	Broadcasted = "webhooks.broadcasted",
+	Failed = "webhooks.failed",
+	Created = "webhooks.created",
+	Updated = "webhooks.updated",
+	Removed = "webhooks.removed",
+}
