--- packages/webhooks/source/events.ts
@@ -1,7 +1,0 @@
-export enum WebhookEvent {
-	Broadcasted = "webhooks.broadcasted",
-	Failed = "webhooks.failed",
-	Created = "webhooks.created",
-	Updated = "webhooks.updated",
-	Removed = "webhooks.removed",
-}
